Let Jit store executable files and subdirectories
Previous versions of jit lacked the ability to manage subdirectories or executable files. This update enables jit to store subdirectories by recursively generating trees. We use depth first traversal to ensure that bottom level trees are written to the database (and thus have a valid oid) before their parents are written. This is necessary as jit utilizes a content addressable database, and the content of the parent trees relies on the oid of the child trees.

We handle executable files by utilizing the stat system call and marking the file's mode at the beginning of each object stored in the database. It is important to note that the mode that jit writes does not directly correspond to the mode returned by stat. Jit is only concerned with whether a file is a directory, is executable (for the current user), or is a regular file. All other special permissions (for example those relating to group read/write permissions) are ignored.