Cleanup debugging statements in commit.rb
The bug was caused by using single quotes to declare a new line 
character. This wrote the character "n" into the contents of my commit
before hashing. The anamolous commit contents led to anamolous hashes
for each commit object which threatened the data integrity of my jit
client.

This was solved by using git rebase, which was able to parse the invalid
commit objects and rewrite their contents sans the "n" character.
Surprisingly, we told git to preserve the commit messages when 
rebasing (that is we did not specify any changes to the commit message).

Rebase accurately extracted the tree, message, author, comitter, and 
I believe even the time of the commit (I could be wrong about the time).
It then re-constructed the commit contents using that extracted data
before rehashing the object. The new (correct) hashes were then used as
contents for the following commit's "parent" field.

In short, git rebase was able to parse my anamolous commit messages,
extract the metadata, rehash the commit, and wrote the new hash under
the parent field of the following commit.

This was all done in a duplicate directory in order to preserve the
original (anamolous) state created by my bug. It is entirely possible
that the rebase had unintended side effects that will show up down the
line, but I am optimistic that the issue was fixed :)