Implement logic to read indexed objects into the cache.
Wrote logic that takes each indexed oid and loads that blob from the database. A DFS traversal of the commit occurs until the index is fully cached. This prepares us for comparing staged changes to the prior commit. New logic is temporarily driven by the show_head script for testing purposes.