Refactor structure of changes data in status command.
Workspace and Index changes are now kept in separate hashes.
This restructuring is to prepare for the --long format output,
which lists workspace and index changes separately.

By keeping separate workspace and index changes in separate and distinct
hashes we are able to keep the code syntactically clear, allowing for easier
maintenance and extensibility of the command.