Fix order of `jit status --long` output
The previous refactor led to a status command that logged changes in the
order they are found by the command's internal logic. Git's 'status'
command logs the files alphabetically by their filename.

To match Git's output a SortedHash class has been implemented in
sorted_hash.rb. @index_changes and @workspace_changes are not instances
of the new SortedHash class, which makes use of a SortedSet to maintain
consistent ordering of the keys when iterating.