Finish implementing the Myer's diff algorithm.
The previous version found the length of the shortest edit path, but not
what that path is. We now push a copy of the v array onto a trace array
during the shortest_edit method. That trace array is used to backtrack
after the finding the shortest path, allowing us to produce a
human-readable diff.