Jit's diff command now shows a line-by-line diff.
Jit's diff can now report differences between the workspace and the
index, as well as differences between the index and HEAD. The diff
algorithm is used to show line-by-line differences.

In its current form, Jit's diff command shows an edit of the entire
file, instead of splitting the edit into hunks as Git normally does.