Pipe the 'diff' output into a pager
The output of the diff command is now piped into a pager. The default
option is less, though other options may be defined through an env
variable. Color formatting is preserved and some flags are passed to the
default pagers (less and lv) to preserve color output.

Additionally there is logic to keep jit running in the background while
the pager process is still active. This is to prevent a zombie pager
process from getting stuck in the background, and allows the user to
continue to interact with the pager via stdin (for navigation and
commands).